PLF-8324 : use offsets option for index_options of file attachment mapping
No explicit mapping were defined for the file attachment content field. The default mapping does not define any value for index_options so the default value 'positions' was applied. This value does not store the start and end character offsets of the words, which are used to speed up highlighting. As explainded in the Elasticsearch, without these offsets, the highlithing process creates a tiny in-memory index and re-runs the original query criteria through Lucene’s query execution planner to get access to low-level match information on the current document. This leads to bad performances when it involves a lot of documents.

This fix updates the file mapping to use the offsets option for index_options and speed up the highlighting process.