Allow to show/hide comments in document preview component configuration
The showComments parameter of the document preview component was not
used, so it was not possible to show/hide comments panel by
configuration.
This PR make use of this parameter so we can now decide to use the
document preview component without the comments feature.
The default value of the showComments parameter is changed to true to
keep the current behavior.

(Rebase: copy commit
https://github.com/exoplatform/platform/commit/5d9174a8d88ad13ace9dcf88a93a03cbdd0e7900)